`grep` for cached repo at the start of the line, not the end
When I went to verify the S3 lookup fix from
14a974b7d0113d0518743b00cb2213de6ca431e9, I noticed new warning
annotations in the Buildkite builds. See
https://github.com/Automattic/git-s3-cache-buildkite-plugin/pull/6#pullrequestreview-1038735114
and
https://buildkite.com/automattic/wordpress-ios/builds/8890#0182006a-2fba-4be6-953f-40a22a7fda9d

After some `echo`-based debugging, I realized the warning was legit in
that the `pre-checkout` code couldn't find the repo in the `manifest`
file from the Git cache server.

The reason for the `grep` failure was that we were looking for the repo
name at the end of the line (`$`) but the `manifest` file format has it
at the start (`^`). See logs in the following build:
https://buildkite.com/automattic/wordpress-ios/builds/8936#01821475-1daf-4f08-91dc-525da9517ae2/231-234

This behavior was introduced in
e95a4f30b248ff948b3b896ae482200edd06fc51, but went unnoticed at the
time.
In all this time, we never noticed because the warning annotation
appeared in the WordPress iOS pipeline only after the work from
14a974b7d0113d0518743b00cb2213de6ca431e9, although I'm not sure why.