Use the first match in the git mirror manifest file
The original implementation assumes there is only one entry
(supposedly the most recent one) for the repo key in the
manifest file—the `URL` would not be an URL if there are
multiple entries. This change removes the assumption and only
greps the first line that starts with the repo key, so that this
plugin would still work even if the manifest file contains multiple
entries for the same key.